+++ src/common/meshcpu.h
@@ -15,12 +15,12 @@ public:
     MeshCPU(std::vector<float> &vertices, std::vector<float> &tex_coords, std::vector<float> &weights)
         : MeshBase()
     {
-        std::vector<float> triangles = BuildTriangles(tex_coords);
+        std::vector<unsigned int> triangles = BuildTriangles(tex_coords);
 
         pos_buffer_ = BufferCPU<float>(vertices);
         tex_buffer_ = BufferCPU<float>(tex_coords);
         wei_buffer_ = BufferCPU<float>(weights);
-        ebo_buffer_ = BufferCPU<float>(triangles);
+        ebo_buffer_ = BufferCPU<unsigned int>(triangles);
 
         pos_size_ = vertices.size();
         tri_size_ = triangles.size();
@@ -56,7 +56,7 @@ private:
     BufferCPU<float> pos_buffer_;
     BufferCPU<float> tex_buffer_;
     BufferCPU<float> wei_buffer_;
-    BufferCPU<float> ebo_buffer_;
+    BufferCPU<unsigned int> ebo_buffer_;
     TextureCPU<float> texture_;
     unsigned int pos_size_;
     unsigned int tri_size_;
